+++ python/paddle/fluid/tests/unittests/dygraph_to_static/test_list.py
@@ -48,6 +48,10 @@ def test_list_in_if(x):
 
 def test_list_in_for_loop(x, iter_num):
     x = fluid.dygraph.to_variable(x)
+    # Use `fill_constant` so that static analysis can analyze the type of iter_num is Tensor
+    iter_num = fluid.layers.fill_constant(
+        shape=[1], value=iter_num, dtype="int32"
+    )  # TODO(liym27): Delete it if the type of parameter iter_num can be resolved
     a = []
     for i in range(iter_num):
         a.append(x)
@@ -57,6 +61,10 @@ def test_list_in_for_loop(x, iter_num):
 def test_list_in_for_loop_with_concat(x, iter_num):
     x = fluid.dygraph.to_variable(x)
     a = []
+    # Use `fill_constant` so that static analysis can analyze the type of iter_num is Tensor
+    iter_num = fluid.layers.fill_constant(
+        shape=[1], value=iter_num, dtype="int32"
+    )  # TODO(liym27): Delete it if the type of parameter iter_num can be resolved
     for i in range(iter_num):
         a.append(x)
     a = fluid.layers.concat(a, axis=0)
